refactor: 썸네일 업로드 로직 서비스 분리 및 책임 정리
