+++ src/domains/course/constants/thumbnail.ts
