+++ src/domains/course/services/resourceService.ts
@@ -21,26 +21,3 @@ export async function createPresignedUploadUrl(
 ): Promise<PresignedUploadUrlResponse> {
   return postApi<PresignedUploadUrlResponse>('/api/instructor/resources', payload);
 }
-
-// Course Thumbnail Presign
-export type CreateCourseThumbnailPresignedUrlRequest = {
-  originalFileName: string;
-  contentType: string;
-  size: number;
-};
-
-export type CreateCourseThumbnailPresignedUrlResponse = {
-  uploadUrl: string;
-  fileKey: string;
-  fileUrl: string;
-  expiresInSeconds: number;
-};
-
-export async function createCourseThumbnailPresignedUrl(
-  payload: CreateCourseThumbnailPresignedUrlRequest,
-) {
-  return postApi<CreateCourseThumbnailPresignedUrlResponse>(
-    '/api/uploads/courses/thumbnails/presign',
-    payload,
-  );
-}
